requirements.txt
langchain==0.1.16
langchain-core==0.1.53
langchain-google-genai==1.0.1
google-generativeai==0.4.1
google-ai-generativelanguage==0.4.0
faiss-cpu
streamlit
python-dotenv
pandas
